refactoring and remove unused references
