+++ src/components/atoms/Button.js
@@ -1,5 +1,4 @@
-import React, { Component } from 'react';
-import { Router, Route, Link, browserHistory } from 'react-router'
+import React from 'react';
 
 const Button = ({text,className,link})=> (
 <Link to={link}>
