+++ src/components/atoms/loginAtoms/InputPassword.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React from 'react';
 
 const InputPassword =({placeholder,password})=> (
 <input type="text"
