+++ src/components/atoms/loginAtoms/InputUser.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React from 'react';
 
 const InputUser =({placeholder})=> (           
   <input type="text"
