+++ src/components/atoms/loginAtoms/Logo.js
@@ -1,7 +1,7 @@
-import React, { Component } from 'react';
+import React from 'react';
 const LoginLogo = ({text, aditional_text}) => (
   <div className="col-md-12">
-    <img src={logoLogin} className="center"></img>
+    <img src={logoLogin} className="l-center"></img>
   </div>
 );
 export default LoginLogo;
