+++ src/components/atoms/userAtoms/UserFullName.js
@@ -1,4 +1,5 @@
 import React from 'react';
+
 const UserFullName = ({text}) => (
   <div className="username">
     {text}
