+++ src/components/atoms/userAtoms/UserImage.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import userImage from '../../images/userImage.png';
+import userImage from '../../../images/userImage.png';
 
 const UserImage = () => (
   <a href="#" className="user-image l-right hidden-xs">
